feat: tip open action
