+++ apps/web/src/components/Common/Publication/PublicationActions.tsx
@@ -24,7 +24,7 @@ type Props = {
 const PublicationActions: FC<Props> = ({ publication }) => {
   const [showTip, setShowTip] = useState(false)
   return (
-    <div className="flex justify-end space-x-1">
+    <div className="mt-4 flex justify-end space-x-1">
       <div className="tape-border flex items-center justify-end overflow-hidden rounded-full bg-gray-100 dark:bg-gray-900">
         <PublicationReaction
           publication={publication}
@@ -60,7 +60,7 @@ const PublicationActions: FC<Props> = ({ publication }) => {
         <OpenActions publication={publication}>
           <div className="flex items-center space-x-1 px-4 py-1 hover:bg-gray-200 dark:hover:bg-gray-800">
             <CollectOutline className="size-4" />
-            <span>Collect</span>
+            <span>Actions</span>
           </div>
         </OpenActions>
       </div>
