+++ apps/web/src/components/Watch/OpenActions/verified-contracts.tsx
@@ -1,0 +1,3 @@
+export const VERIFIED_UNKNOWN_OPEN_ACTION_CONTRACTS = {
+  TIP: '0x22cb67432C101a9b6fE0F9ab542c8ADD5DD48153'
+}
