+++ packages/ui/index.ts
@@ -12,6 +12,7 @@ export * from './src/players/audio'
 export * from './src/players/video'
 export * from './src/Popover'
 export * from './src/QRCode'
+export * from './src/RangeSlider'
 export * from './src/Select'
 export * from './src/Spinner'
 export * from './src/Switch'
