+++ packages/ui/package.json
@@ -21,10 +21,11 @@
     "@radix-ui/react-hover-card": "^1.0.7",
     "@radix-ui/react-popover": "^1.0.7",
     "@radix-ui/react-select": "^2.0.0",
+    "@radix-ui/react-slider": "^1.1.2",
     "@radix-ui/react-switch": "^1.0.3",
     "@radix-ui/react-tabs": "^1.0.4",
-    "clsx": "^2.1.0",
     "@tippyjs/react": "^4.2.6",
+    "clsx": "^2.1.0",
     "wavesurfer.js": "^7.6.3"
   }
 }
